add method verification

+++ internal/router/endpoints/posts.go
@@ -14,22 +14,28 @@ func NewPostHandler(w http.ResponseWriter, r *http.Request) {
 	var R models.HttpResponse
 	var code int
 
-	err := utils.GetPostRequestData(r, &post)
-	if err != nil {
-		code = http.StatusUnprocessableEntity
-		R.Status = "failure"
-		R.Message = "unable to create post object"
-	} else {
-		err = database.NewPost(post)
+	if utils.VerifyMethod(r, "POST") {
+		err := utils.GetPostRequestData(r, &post)
 		if err != nil {
-			code = http.StatusInternalServerError
+			code = http.StatusUnprocessableEntity
 			R.Status = "failure"
-			R.Message = "unable to save post object"
+			R.Message = "unable to create post object"
 		} else {
-			code = http.StatusCreated
-			R.Status = "success"
-			R.Message = "new post created"
+			err = database.NewPost(post)
+			if err != nil {
+				code = http.StatusInternalServerError
+				R.Status = "failure"
+				R.Message = "unable to save post object"
+			} else {
+				code = http.StatusCreated
+				R.Status = "success"
+				R.Message = "new post created"
+			}
 		}
+	} else {
+		code = 405
+		R.Status = "failure"
+		R.Message = "method not allowed"
 	}
 	json.NewEncoder(w).Encode(R)
 	w.WriteHeader(code)
@@ -41,29 +47,36 @@ func DeletePostHandler(w http.ResponseWriter, r *http.Request) {
 
 	var R models.HttpResponse
 	var code int
-	if !ok || len(args[0]) < 1 {
-		code = http.StatusNotAcceptable
-		R.Status = "failure"
-		R.Message = "no ID value provided"
-	} else {
-		var err error
-		id, err = strconv.Atoi(args[0])
-		if err != nil {
-			code = http.StatusUnprocessableEntity
+
+	if utils.VerifyMethod(r, "POST") {
+		if !ok || len(args[0]) < 1 {
+			code = http.StatusNotAcceptable
 			R.Status = "failure"
-			R.Message = err.Error()
+			R.Message = "no ID value provided"
 		} else {
-			err = database.DeletePost(id)
+			var err error
+			id, err = strconv.Atoi(args[0])
 			if err != nil {
-				code = http.StatusInternalServerError
+				code = http.StatusUnprocessableEntity
 				R.Status = "failure"
 				R.Message = err.Error()
 			} else {
-				code = http.StatusOK
-				R.Status = "success"
-				R.Message = "post deleted"
+				err = database.DeletePost(id)
+				if err != nil {
+					code = http.StatusInternalServerError
+					R.Status = "failure"
+					R.Message = err.Error()
+				} else {
+					code = http.StatusOK
+					R.Status = "success"
+					R.Message = "post deleted"
+				}
 			}
 		}
+	} else {
+		code = 405
+		R.Status = "failure"
+		R.Message = "method not allowed"
 	}
 	json.NewEncoder(w).Encode(R)
 	w.WriteHeader(code)
@@ -75,42 +88,48 @@ func GetPostHandler(w http.ResponseWriter, r *http.Request) {
 
 	var R models.HttpResponse
 	var code int
-	if !ok || len(args[0]) < 1 {
-		code = http.StatusNotAcceptable
-		R.Status = "failure"
-		R.Message = "no ID value provided"
-	} else {
-		var err error
-		id, err = strconv.Atoi(args[0])
-		if err != nil {
-			code = http.StatusUnprocessableEntity
+
+	if utils.VerifyMethod(r, "GET") {
+		if !ok || len(args[0]) < 1 {
+			code = http.StatusNotAcceptable
 			R.Status = "failure"
-			R.Message = err.Error()
+			R.Message = "no ID value provided"
 		} else {
-			post, err := database.GetPost(id)
+			var err error
+			id, err = strconv.Atoi(args[0])
 			if err != nil {
-				code = http.StatusInternalServerError
+				code = http.StatusUnprocessableEntity
 				R.Status = "failure"
 				R.Message = err.Error()
-			} else if post.Id == 0 {
-				code = http.StatusNotFound
-				R.Status = "failure"
-				R.Message = "post not found"
 			} else {
-				post, err := json.Marshal(post)
+				post, err := database.GetPost(id)
 				if err != nil {
 					code = http.StatusInternalServerError
 					R.Status = "failure"
-					R.Message = "unable to create object"
+					R.Message = err.Error()
+				} else if post.Id == 0 {
+					code = http.StatusNotFound
+					R.Status = "failure"
+					R.Message = "post not found"
 				} else {
-					code = http.StatusInternalServerError
-					R.Status = "success"
-					R.Message = "post retrieved"
-					R.Payload = json.RawMessage(post)
+					post, err := json.Marshal(post)
+					if err != nil {
+						code = http.StatusInternalServerError
+						R.Status = "failure"
+						R.Message = "unable to create object"
+					} else {
+						code = http.StatusInternalServerError
+						R.Status = "success"
+						R.Message = "post retrieved"
+						R.Payload = json.RawMessage(post)
+					}
 				}
-
 			}
 		}
+	} else {
+		code = 405
+		R.Status = "failure"
+		R.Message = "method not allowed"
 	}
 	json.NewEncoder(w).Encode(R)
 	w.WriteHeader(code)
@@ -123,26 +142,31 @@ func AllPostsHandler(w http.ResponseWriter, r *http.Request) {
 	var R models.HttpResponse
 	var code int
 
-	if !ok || len(args[0]) < 1 {
-		limit = 0
-	} else {
-		var err error
-		limit, err = strconv.Atoi(args[0])
-		if err != nil {
+	if utils.VerifyMethod(r, "GET") {
+		if !ok || len(args[0]) < 1 {
 			limit = 0
+		} else {
+			var err error
+			limit, err = strconv.Atoi(args[0])
+			if err != nil {
+				limit = 0
+			}
 		}
-	}
-
-	_, posts, err := database.GetPosts(limit)
-	if err != nil {
-		code = http.StatusInternalServerError
-		R.Status = "failure"
-		R.Message = "unable to get posts"
+		_, posts, err := database.GetPosts(limit)
+		if err != nil {
+			code = http.StatusInternalServerError
+			R.Status = "failure"
+			R.Message = "unable to get posts"
+		} else {
+			code = http.StatusOK
+			R.Status = "success"
+			R.Message = "posts retrieved"
+			R.Payload = posts
+		}
 	} else {
-		code = http.StatusOK
-		R.Status = "success"
-		R.Message = "posts retrieved"
-		R.Payload = posts
+		code = 405
+		R.Status = "failure"
+		R.Message = "method not allowed"
 	}
 	json.NewEncoder(w).Encode(R)
 	w.WriteHeader(code)
